fix(build): keep node:crypto out of the client bundle (newsletter A/B split)
main deploy has been red since ~08:50 (every commit since #2075 fails the
`build` job). The surfaced error is misleading:

  [og-pages] [spa-bundle-resolver] poll exhausted ... index.html does not exist yet

That is a RED HERRING. The real, masked failure is a fatal rollup error during
the client build:

  services/newsletter-subject-variants.mjs (24:9): "createHash" is not exported
  by "__vite-browser-external", imported by services/newsletter-subject-variants.mjs

Chain: the SPA graph reaches this module via
services/newsletterPreview.ts → services/newsletter-content.mjs →
getVariantStyleDirective (newsletter-subject-variants.mjs). PR #2075 added that
`getVariantStyleDirective` import to newsletter-content.mjs, and the variants
module statically imports `node:crypto`'s `createHash` (used by
`assignSubjectVariant`). In the browser build `node:crypto` is stubbed as
`__vite-browser-external`, which does not export `createHash`, so rollup aborts
the client render. index.html/index-entry.js are never emitted, the render phase
is skipped entirely, and og-pages' closeBundle (run during teardown) then throws
its index.html poll-exhaustion error — hiding the real cause. #2075's own deploy
was cancelled (superseded), so the failure first appeared one commit later at
#2085, which is why it looked like the salary-stats PR.

Fix: move the only crypto consumer, `assignSubjectVariant`, into a new
server-only module `services/newsletter-subject-assign.mjs`, so the
client-reachable `newsletter-subject-variants.mjs` no longer imports any `node:`
builtin. The hash is byte-identical (same `createHash('sha256')` → uint32 mod),
so variant assignment is unchanged and the A/B report can still re-compute
historical variants — no re-bucketing. The two server scripts (send-newsletter,
newsletter-ab-report) and the unit test import `assignSubjectVariant` from the
new module.

Verified locally on origin/main: before the fix `vite build --minify esbuild`
errors on `createHash` and emits no index.html; after, it renders chunks and
writes dist/index.html (108 KB) + dist/assets/index-entry.js. Unit suite
tests/newsletter-subject-variants.test.ts: 27/27 pass.

Closes #2024
